feat: add usePayrollTableDialog for managing payroll dialog state and actions
